Fix iOS build: change seed to UInt32 everywhere to match Stepper range
MLXSettings.seed: UInt64 -> UInt32 (matches 0...UInt32.max Stepper)
BackgroundTrainingConfig.seed: UInt64 -> UInt32 (consistent with MLXSettings)
MLXDeviceLoopController: cast UInt32 seed to UInt64 for MLXLoopConfig
BackgroundTaskScheduler: cast UInt32 config.seed to UInt64 for MLXLoopConfig

+++ App/SwairmApp/BackgroundTaskScheduler.swift
@@ -53,7 +53,7 @@ struct BackgroundTrainingConfig: Codable, Sendable {
     var targetModules: String
     var loraRank: Int
     var loraAlpha: Float
-    var seed: UInt64
+    var seed: UInt32
     var curriculumEpoch: Int
     var minBatteryFraction: Float?
     var hmacSecret: Data?
@@ -410,7 +410,7 @@ final class BackgroundTaskScheduler {
                         batchSize: config.batchSize,
                         sequenceLength: config.sequenceLength,
                         curriculumDirectory: resolvedInDocuments("curriculum/epoch_\(config.curriculumEpoch)"),
-                        seed: config.seed + UInt64(config.deviceIndex)
+                        seed: UInt64(config.seed) + UInt64(config.deviceIndex)
                     )
                 )
